[WIP] Re-add disabled tests
